ai-server/requirements.txt
fastapi
uvicorn
langchain-openai>=0.1.0
pydantic
python-dotenv